Squashed commit of the following:

commit 26550d4f1bc7f35fdf153cd46a9466820c7620c7
Author: huica <huica@cisco.com>
Date:   Wed Feb 21 14:39:57 2018 -0500

    File api: make sure file instance is unique accross sessions

+++ src/file_api/file_enforcer.cc
@@ -96,7 +96,7 @@ FileVerdict FileEnforcer::check_verdict(Flow* flow, FileNode* node,
 int FileEnforcer::store_verdict(Flow* flow, FileInfo* file)
 {
     assert(file);
-    size_t file_sig = file->get_file_id();
+    uint64_t file_sig = file->get_file_id();
 
     if (!file_sig)
         return 0;
@@ -215,7 +215,7 @@ FileVerdict FileEnforcer::cached_verdict_lookup(Flow* flow, FileInfo* file,
     }
 
     assert(file);
-    size_t file_sig = file->get_file_id();
+    uint64_t file_sig = file->get_file_id();
     if (!file_sig)
         return verdict;
